requirements.txt
django
pillow
psycopg2-binary
python-dotenv
djangorestframework
djangorestframework-simplejwt
drf-yasg
django-cors-headers
celery
eventlet
redis
coverage